Moved system language analyzers to separate files.
