+++ src/Compilers/Core/CodeAnalysisTest/CodeAnalysisTest.csproj
@@ -29,9 +29,10 @@
     <Compile Include="Collections\ByteSequenceComparerTests.cs" />
     <Compile Include="Diagnostics\BoxingOperationAnalyzer.cs" />
     <Compile Include="Diagnostics\DiagnosticLocalizationTests.cs" />
-    <Compile Include="Diagnostics\EmptyArrayOperationAnalyzer.cs" />
+    <Compile Include="Diagnostics\EmptyArrayAnalyzer.cs" />
+    <Compile Include="Diagnostics\LocalCouldBeConstAnalyzer.cs" />
     <Compile Include="Diagnostics\OperationTestAnalyzer.cs" />
-    <Compile Include="Diagnostics\SystemLanguageAnalyzer.cs" />
+    <Compile Include="Diagnostics\FieldCouldBeReadOnlyAnalyzer.cs" />
     <Compile Include="Emit\EmitOptionsTests.cs" />
     <Compile Include="Emit\CustomDebugInfoTests.cs" />
     <Compile Include="InternalUtilities\StreamExtensionsTests.cs" />
